backend/requirements.txt
django
djangorestframework
django-cors-headers
python-dotenv
Pillow
dj-rest-auth
django-allauth
drf-spectacular
psycopg2-binary
reportlab
pypdf
